refactor: move user entity for domain
